+++ src/experiments/prompty.py
@@ -1,0 +1,6 @@
+from src.experiments.base_experiment import Experiment
+
+
+class PromptyExperiment(Experiment):
+    def create(self):
+        raise NotImplementedError("Prompty experiment not implemented yet.")
