+++ src/experiments/pure_python.py
@@ -1,0 +1,6 @@
+from src.experiments.base_experiment import Experiment
+
+
+class PythonExperiment(Experiment):
+    def create(self):
+        raise NotImplementedError("Python experiment not implemented yet.")
